# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FraudGuard API
  version: "1.0.0"
  description: |
    REST API for FraudGuard — a hybrid rules + ML transaction fraud-detection
    system. Every `POST /transactions` call runs the submitted transaction
    through the real hybrid engine (hard rules → ML model → rule-engine
    fallback) and returns a `risk_level`, `fraud_score`, human-readable
    `fraud_reasons`, and — when the ML model scored it — a SHAP-based
    `shap_explanation` showing which features drove the score.

    See `docs/METHODOLOGY.md` for how detection works and the trade-offs
    behind it, and the main `README.md` for an architecture diagram and
    setup instructions.

    **Versioning:** all routes are served under `/api/v1`. The unversioned
    `/api/*` paths shown here also work today (mounted as an alias to the
    same v1 routes) but new integrations should target `/api/v1` — see
    the `servers` list below.
  contact:
    name: FraudGuard
  license:
    name: UNLICENSED
servers:
  - url: http://localhost:5000/api/v1
    description: Local development (versioned — preferred)
  - url: http://localhost:5000/api
    description: Local development (unversioned alias)
...
security:
  - bearerAuth: []
paths:
...
  /alerts:
    get:
      tags: [Alerts]
      summary: List/search alerts (paginated, filterable)
      description: |
        Regular users see only their own alerts; admins see every user's.
        All filter params are optional and combinable — this is the
        case-management filter bar behind the Alerts page.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - name: status
          in: query
          schema: { type: string, enum: [open, in_review, resolved] }
        - name: riskLevel
          in: query
          schema: { type: string, enum: [medium, high] }
        - name: merchant
          in: query
          description: Case-insensitive partial match on the underlying transaction's merchant.
          schema: { type: string }
        - name: dateFrom
          in: query
          schema: { type: string, format: date-time }
        - name: dateTo
          in: query
          schema: { type: string, format: date-time }
        - name: amountMin
          in: query
          schema: { type: number }
        - name: amountMax
          in: query
          schema: { type: number }
        - name: assignedTo
          in: query
          description: Filter to alerts assigned to a specific user id.
          schema: { type: integer }
      responses:
        "200":
          description: A page of alerts
          content:
            application/json:
              schema:
                type: object
                properties:
                  alerts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Alert"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "400":
          description: Invalid filter value (e.g. unrecognized status)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /alerts/{id}:
    get:
      tags: [Alerts]
      summary: View a single alert
      description: Audit-logged (`alerts.view`). Only the alert's owner or an admin may view it.
      parameters:
        - $ref: "#/components/parameters/AlertId"
      responses:
        "200":
          description: The alert
          content:
            application/json:
              schema:
                type: object
                properties:
                  alert:
                    $ref: "#/components/schemas/Alert"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /alerts/{id}/resolve:
    patch:
      tags: [Alerts]
      summary: Resolve an alert with an analyst verdict
      description: |
        Audit-logged (`alerts.resolve`, recording the verdict and whether
        a note was given). Only the alert's owner or an admin may resolve
        it. `verdict` is required, not optional — this is what
        `transactionRepository.fraudDeviceAndIpBlacklist` reads to decide
        whether the transaction's device/IP belongs on the dynamic
        blacklist: `confirmed_fraud` adds it regardless of the model's
        original score, and `false_positive` explicitly removes it even
        if the model originally flagged it, closing the analyst feedback
        loop instead of leaving that decision purely model-driven.
      parameters:
        - $ref: "#/components/parameters/AlertId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveAlertRequest"
            example:
              verdict: confirmed_fraud
              note: Cardholder confirmed the card was reported stolen.
      responses:
        "200":
          description: Alert resolved
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: Alert resolved }
                  alert:
                    $ref: "#/components/schemas/Alert"
        "400":
          description: Missing/invalid verdict, or note too long
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /admin/audit-logs/alert/{id}:
    get:
      tags: [Admin]
      summary: Full view/resolve history for one alert
      parameters:
        - $ref: "#/components/parameters/AlertId"
      responses:
        "200":
          description: Every audit entry targeting this alert
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      $ref: "#/components/schemas/AuditLogEntry"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api-keys:
    post:
      tags: [API Keys]
      summary: Create an API key for your own account
      description: |
        Self-service — any logged-in user can create keys for their own
        account (the same ownership model as webhooks below). Returns
        the raw key in plaintext exactly once; only its SHA-256 hash is
        ever stored server-side.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 100, example: "Production Backend" }
                scopes:
                  type: array
                  items: { type: string, enum: [transactions:write, transactions:read] }
                  default: [transactions:write, transactions:read]
                expiresInDays:
                  type: integer
                  minimum: 1
                  maximum: 3650
                  description: Omit for no expiration (or the server-configured default — API_KEY_DEFAULT_EXPIRES_DAYS).
      responses:
        "201":
          description: API key created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }
                  name: { type: string }
                  key: { type: string, description: "Shown only in this response — cannot be retrieved again.", example: "fg_live_9f8e7d6c5b4a3928" }
                  prefix: { type: string, example: "fg_live_9f8e" }
                  scopes:
                    type: array
                    items: { type: string }
                  expires_at: { type: string, format: date-time, nullable: true }
                  created_at: { type: string, format: date-time }
                  warning: { type: string }
        "400":
          description: Invalid name, scopes, or expiresInDays
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    get:
      tags: [API Keys]
      summary: List your own API keys
      description: |
        Never includes the key value or its hash. Admins see every
        user's keys by default — pass `mine=true` to scope the list to
        just the caller's own even as an admin (used by the self-service
        Settings page so it never mixes in other users' keys).
      parameters:
        - name: mine
          in: query
          schema: { type: string, enum: ["true"] }
          description: Forces the list to just the caller's own keys, even for an admin.
      responses:
        "200":
          description: API keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  apiKeys:
                    type: array
                    items:
                      $ref: "#/components/schemas/ApiKey"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api-keys/{id}:
    delete:
      tags: [API Keys]
      summary: Revoke an API key
      description: Owner or admin. Idempotent — revoking an already-revoked key returns 200 without error.
      parameters:
        - $ref: "#/components/parameters/ApiKeyId"
      responses:
        "200":
          description: Revoked (or already was)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Access token returned by `/auth/login`, `/auth/register`, or `/auth/refresh`. Send as `Authorization: Bearer <accessToken>`."
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Service-to-service auth (feature: API keys) — an alternative to
        a JWT session for integrations that shouldn't hold a human's
        password-derived credential. Create one via `POST /api-keys`
        (shown in plaintext exactly once). Acts as the account that
        created it, subject to the key's own `scopes`
        (`transactions:read`/`transactions:write`) — see
        `middleware/apiKeyAuth.js`. Only accepted on the
        `/transactions*` routes; alerts/admin stay JWT-only.
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    AlertId:
      name: id
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
    ApiKeyId:
      name: id
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
...
  responses:
    Unauthorized:
      description: Missing, invalid, or expired access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated, but not allowed to perform this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
        requestId: { type: string, description: "Echoed from the X-Request-Id response header — include this when reporting an issue." }
    ValidationErrorResponse:
      type: object
      properties:
        error: { type: string, example: Validation failed }
        details:
          type: array
          items:
            type: object
            properties:
              field: { type: string, example: amount }
              message: { type: string, example: Amount must be greater than 0 }
...
    Alert:
      type: object
      properties:
        id: { type: integer }
        transaction_id: { type: integer }
        user_id: { type: integer }
        message: { type: string }
        risk_level: { type: string, enum: [medium, high] }
        resolved: { type: integer, enum: [0, 1], description: "Legacy boolean, kept in sync with status = 'resolved'. Prefer status for new integrations." }
        status: { type: string, enum: [open, in_review, resolved] }
        verdict: { type: string, enum: [confirmed_fraud, false_positive], nullable: true, description: "Set only once resolved — see PATCH /alerts/{id}/resolve." }
        resolution_note: { type: string, nullable: true }
        assigned_to: { type: integer, nullable: true, description: "User id of the assigned analyst, if any." }
        assignee_username: { type: string, nullable: true }
        resolved_by: { type: integer, nullable: true }
        resolved_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        merchant: { type: string }
        amount: { type: number, format: float }
        username: { type: string, description: "Present on admin's alerts.allWithUsername view" }
    ResolveAlertRequest:
      type: object
      required: [verdict]
      properties:
        verdict:
          type: string
          enum: [confirmed_fraud, false_positive]
          description: Required — the whole point of this endpoint is recording an analyst's confirmed/false-positive call, not just flipping a boolean.
        note:
          type: string
          maxLength: 1000
...
    Pagination:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        total_pages: { type: integer }
        has_next: { type: boolean }
        has_prev: { type: boolean }
...
    AuditLogEntry:
      type: object
      properties:
        id: { type: integer }
        created_at: { type: string, format: date-time }
        user_id: { type: integer, nullable: true }
        username: { type: string, nullable: true }
        action: { type: string, example: alerts.resolve }
        target_type: { type: string, nullable: true, example: alert }
        target_id: { type: integer, nullable: true }
        outcome: { type: string, enum: [success, failure, denied] }
        details: { type: object, additionalProperties: true }
    ApiKey:
      type: object
      description: Never includes the key value or its hash — see POST /api-keys for the one-time plaintext response.
      properties:
        id: { type: integer }
        user_id: { type: integer }
        username: { type: string, description: "Present on admin's list view" }
        name: { type: string }
        key_prefix: { type: string, example: "fg_live_9f8e" }
        scopes:
          type: array
          items: { type: string, enum: [transactions:write, transactions:read] }
        created_by: { type: integer, nullable: true }
        expires_at: { type: string, format: date-time, nullable: true, description: "null means the key never expires." }
        last_used_at: { type: string, format: date-time, nullable: true }
        revoked_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
...